requirements.txt
opencv-python
torch
torchvision
torchaudio
ultralytics
pillow
google-generativeai
SpeechRecognition
pyttsx3
pyaudio
requests
fastapi
uvicorn
python-multipart
python-dotenv
numpy
scipy
timm